[TM-882] The abort/assert pattern is useful, bump it out into a trait.
